requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-decouple==3.8
drf-spectacular==0.26.3